EES-4577 Move server validation error handling from `RHFForm` to `FormProvider`
This change moves the server validation error handling to `FormProvider` so that
we can integrate it into any of the RHF `handleSubmit` calls.

Previously, this handling was only being done in `RHFForm`, meaning that it
would only trigger during `form` submit events. Quite a few of our forms delay
form submission until a confirmation button is clicked, meaning that the error
handling would not kick in for those cases.
