+++ src/explore-education-statistics-common/src/components/form/rhf/FormProvider.tsx
@@ -1,11 +1,19 @@
+import { Path } from '@common/types';
+import {
+  FieldMessageMapper,
+  isServerValidationError,
+  mapServerFieldErrors,
+} from '@common/validation/serverValidations';
+import has from 'lodash/has';
 import {
   FieldValues,
   FormProvider as RHFormProvider,
   useForm,
+  UseFormHandleSubmit,
   UseFormProps,
   UseFormReturn,
 } from 'react-hook-form';
-import React, { ReactNode, useEffect, useRef } from 'react';
+import React, { ReactNode, useCallback, useEffect, useRef } from 'react';
 import { yupResolver } from '@hookform/resolvers/yup';
 import isEqual from 'lodash/isEqual';
 import { ObjectSchema, Schema } from 'yup';
@@ -13,6 +21,10 @@ import { ObjectSchema, Schema } from 'yup';
 interface FormProviderProps<TFormValues extends FieldValues> {
   children: ReactNode | ((form: UseFormReturn<TFormValues>) => ReactNode);
   enableReinitialize?: boolean;
+  errorMappers?:
+    | FieldMessageMapper<TFormValues>[]
+    | ((values: TFormValues) => FieldMessageMapper<TFormValues>[]);
+  fallbackErrorMapper?: FieldMessageMapper<TFormValues>;
   initialValues?: UseFormProps<TFormValues>['defaultValues'];
   validationSchema?: ObjectSchema<TFormValues> & Schema<TFormValues>;
 }
@@ -20,6 +32,8 @@ interface FormProviderProps<TFormValues extends FieldValues> {
 export default function FormProvider<TFormValues extends FieldValues>({
   children,
   enableReinitialize,
+  errorMappers = [],
+  fallbackErrorMapper,
   initialValues,
   validationSchema,
 }: FormProviderProps<TFormValues>) {
@@ -33,11 +47,11 @@ export default function FormProvider<TFormValues extends FieldValues>({
   const previousInitialValues = useRef(initialValues);
 
   /**
-   * RHF caches default values and doesn't have a built in option
+   * RHF caches default values and doesn't have a built-in option
    * to reinitialise forms.
    * This use effect provides the equivalent functionality as
    * Formik's `enableReinitialize` by checking for changes to the
-   * initialValues and reseting the form with the new values if they've changed.
+   * initialValues and resetting the form with the new values if they've changed.
    */
   useEffect(() => {
     if (
@@ -49,9 +63,45 @@ export default function FormProvider<TFormValues extends FieldValues>({
     }
   }, [enableReinitialize, form, initialValues]);
 
+  const handleSubmit: UseFormHandleSubmit<TFormValues> = useCallback(
+    (onValid, onInvalid) => {
+      return form.handleSubmit(async (values, event) => {
+        try {
+          await onValid(values, event);
+        } catch (error) {
+          if (isServerValidationError(error) && error.response?.data) {
+            const fieldErrors = mapServerFieldErrors(
+              error.response.data,
+              typeof errorMappers === 'function'
+                ? errorMappers(values as TFormValues)
+                : errorMappers,
+              fallbackErrorMapper,
+            );
+
+            fieldErrors.forEach(({ field, message }) => {
+              if (has(values, field)) {
+                form.setError(field as Path<TFormValues>, { message });
+              }
+            });
+
+            return;
+          }
+
+          throw error;
+        }
+      }, onInvalid);
+    },
+    [errorMappers, fallbackErrorMapper, form],
+  );
+
+  const providerProps = {
+    ...form,
+    handleSubmit,
+  };
+
   return (
-    <RHFormProvider {...form}>
-      {typeof children === 'function' ? children(form) : children}
+    <RHFormProvider {...providerProps}>
+      {typeof children === 'function' ? children(providerProps) : children}
     </RHFormProvider>
   );
 }
